Wellness_app/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
google-generativeai==0.3.1 